requirements.txt
pandas
scikit-learn
numpy
flask
python-dotenv